Move all http requests related to books to router/books.js
